AMPUI-44 migrate funding sources- remove former aditionalInfo. fix eslint
